chore(amazonq): move nep pojo to model.ts
